requirements.txt
Pillow==6.0.0
numpy==1.16.3
torch==1.1.0
torchvision
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16
scikit-learn==0.21.1
sklearn==0.0
jupyter==1.0.0